chore: 카테고리 페이지 및 해당 페이지의 API함수 TS로 변환
