--- 빵세-방세현/3_7주차/MakeMovieSite/src/hooks/queries/getMovies.js
@@ -1,11 +1,0 @@
-import { axiosInstance } from "../../apis/axios-instance";
-
-const getMovies = async ({ category, pageParam }) => {
-  const { data } = await axiosInstance.get(
-    `/movie/${category}?language=ko&page=${pageParam}`
-  );
-
-  return data;
-};
-
-export { getMovies };
